feat(podcast): set value object form

+++ src/Domain/Shared/Entity/ThumbnailTrait.php
@@ -18,7 +18,7 @@ trait ThumbnailTrait
 
     private ?string $thumbnail_url = null;
 
-    private ?float $thumbnail_size = null;
+    private ?int $thumbnail_size = null;
 
     private ?array $thumbnail_dimensions = [];
